Move all executor based tests to one module
