+++ tests/test_pylammpsmpi_local.py
@@ -6,9 +6,8 @@ import sys
 import unittest
 
 import numpy as np
-from executorlib import SingleNodeExecutor
 
-from pylammpsmpi import LammpsLibrary, init_function
+from pylammpsmpi import LammpsLibrary
 
 
 class TestLocalLammpsLibrary(unittest.TestCase):
@@ -134,39 +133,6 @@ class TestLocalLammpsLibrary(unittest.TestCase):
     def test_create_atoms_typeerror(self):
         with self.assertRaises(TypeError):
             self.lmp.create_atoms(n=1, atomid=[1], atype=[1], x=None)
-
-
-class TestExecutorLammpsLibrary(unittest.TestCase):
-    def test_version(self):
-        execution_path = os.path.dirname(os.path.abspath(__file__))
-        citation_file = os.path.join(execution_path, "citations.txt")
-        lammps_file = os.path.join(execution_path, "in.simple")
-        with SingleNodeExecutor(
-            block_allocation=True,
-            hostname_localhost=True,
-            max_workers=1,
-            init_function=init_function,
-            openmpi_oversubscribe=False,
-            resource_dict={
-                "cores": 2,
-                "cwd": ".",
-            },
-        ) as exe:
-            lmp = LammpsLibrary(cores=2, cmdargs=["-cite", citation_file], executor=exe)
-            lmp.file(lammps_file)
-            self.assertTrue(
-                lmp.version
-                in [
-                    20220623,
-                    20230802,
-                    20231121,
-                    20240207,
-                    20240627,
-                    20240829,
-                    20250722,
-                ]
-            )
-            lmp.close()
 
 
 if __name__ == "__main__":
